pyproject.toml
[tool.black]
skip-string-normalization = true

[tool.pytest.ini_options]
testpaths = ["mpl_uncertainties/tests"]

[build-system]
requires = ["setuptools>=42", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "mpl-uncertainties"
dynamic = ["version"]
readme = "README.md"
dependencies = ["uncertainties>=3.2.3,<4.0.0", "matplotlib", "scipy"]

[dependency-groups]
tests = ["pytest", "pytest-mpl"]
docs = [
  "sphinx",
  "sphinx_copybutton",
  "sphinx_gallery",
  "numpydoc",
  "autoapi",
  "scipy",
  "furo",
]

[tool.pixi.project]
channels = ["conda-forge"]
platforms = ["win-64"]

[tool.pixi.pypi-dependencies]
mpl-uncertainties = { path = ".", editable = true }

[tool.pixi.tasks]

[tool.pixi.feature.tests.tasks]
test = "pytest --mpl --mpl-baseline-path=baseline --mpl-default-style=default"
test_mpl_generate = "pytest --mpl-generate-path=baseline --mpl-default-style=default"

[tool.pixi.environments]
default = { solve-group = "default" }
tests = { features = ["tests"], solve-group = "default" }
docs = { features = ["docs"], solve-group = "default" }
lint = { features = ["lint"], solve-group = "default" }

[tool.pixi.feature.docs.tasks]
build-doc = { cmd = "make html", cwd = "docs" }
clean-doc = { cmd = "rm -rf _build examples", cwd = "docs" }
example = { cmd = "python examples/errorbar.py" }

# doc gets used by readthedocs, docs is used by pixi
# test gets used by pip/github actions, tests is used by pixi
[project.optional-dependencies]
doc = [
  "sphinx",
  "sphinx_copybutton",
  "sphinx_gallery",
  "numpydoc",
  "autoapi",
  "scipy",
  "furo",
]
test = ["pytest", "pytest-mpl"]

[tool.pixi.feature.lint.dependencies]
pre-commit = "*"
pre-commit-hooks = "*"
taplo = "*"
ruff = "*"
mdformat = "*"
mdformat-ruff = "*"

[tool.pixi.feature.lint.tasks]
pre-commit-install = "pre-commit install"
lint = "pre-commit run --all-files"
